backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
psycopg[binary]==3.2.6
python-dotenv==1.0.1
Pillow==11.2.1
